Move Mojo VARIANT cross-call branch tests to tests/integration
Coverage for the homogeneous-slot short-circuit and the unmappable-
scalar fallback now lives in `tests/integration/`, alongside the other
`literalize_call` integration tests.
