+++ tests/test_languages.py
@@ -27,7 +27,6 @@ from literalizer.languages import (
     Java,
     Jsonnet,
     Matlab,
-    Mojo,
     Python,
     R,
     Rust,
@@ -471,65 +470,6 @@ def test_gleam_special_floats_raise(yaml_value: str) -> None:
             input_format=InputFormat.YAML,
             language=Gleam(),
         )
-
-
-def test_mojo_variant_call_slot_with_unmappable_scalar_falls_back() -> None:
-    """A Mojo VARIANT call slot containing an unmappable scalar (e.g.
-    ``None``) falls back to the generic ``[*Ts: AnyType]`` stub.
-
-    ``None`` is not in Mojo's call-arg scalar mapping, so
-    ``_value_to_mojo_type`` returns ``None`` for it.  The typed-param
-    helper detects the unresolved slot type and bails to the generic
-    signature instead of emitting ``Variant[...]``.  This exercises the
-    ``None in slot_types`` fallback branch.
-    """
-    source = "---\n- null\n- hello\n"
-    variant_strategy = next(
-        strategy
-        for strategy in Mojo.heterogeneous_strategies
-        if strategy.name == "VARIANT"
-    )
-    spec = Mojo(heterogeneous_strategy=variant_strategy)
-    result = literalize_call(
-        source=source,
-        input_format=InputFormat.YAML,
-        language=spec,
-        target_function="process",
-        parameter_names=["value"],
-        per_element=True,
-        wrap_in_file=True,
-    )
-    assert "def process[*Ts: AnyType](*args: *Ts):" in result.code
-
-
-def test_mojo_variant_homogeneous_slot_skipped_in_wrap_ids() -> None:
-    """A homogeneous all-scalar VARIANT slot contributes no wrap ids.
-
-    With a 2-param Mojo VARIANT call where slot 0 diverges
-    (``String, Int, Bool``) but slot 1 is uniformly ``String``, the
-    cross-call wrap-id helper wraps slot-0 values but leaves slot-1
-    values untouched.  This exercises the ``len(slot_types) <= 1``
-    short-circuit in ``_mojo_cross_call_scalar_wrap_ids``.
-    """
-    source = "---\n- - hello\n  - a\n- - 42\n  - b\n- - true\n  - c\n"
-    variant_strategy = next(
-        strategy
-        for strategy in Mojo.heterogeneous_strategies
-        if strategy.name == "VARIANT"
-    )
-    spec = Mojo(heterogeneous_strategy=variant_strategy)
-    result = literalize_call(
-        source=source,
-        input_format=InputFormat.YAML,
-        language=spec,
-        target_function="process",
-        parameter_names=["divergent", "uniform"],
-        per_element=True,
-        wrap_in_file=True,
-    )
-    assert 'process(Value(String("hello")), "a")' in result.code
-    assert 'process(Value(42), "b")' in result.code
-    assert 'process(Value(True), "c")' in result.code
 
 
 def test_jsonnet_wrap_calls_with_declarations_prepends_bindings() -> None:
